edit-tab-again * Rename RealtimeDataPanel directory to RealtimeDetailPanel
