+++ src/containers/Admin/EditTab/RealtimeTile/RealtimeDetailPanel/PermanentLinePanel.module.scss
