+++ src/containers/Admin/EditTab/RealtimeTile/RealtimeDetailPanel/RealtimeDetailPanel.module.scss
